Issue #68: Test restructured and passing
