+++ datastore/db/queries/inference/__init__.py
@@ -107,7 +107,8 @@ def set_inference_feedback_user_id(cursor, inference_id, user_id):
                 id = %s
             """
         cursor.execute(query, (user_id,inference_id))
-    except Exception:
+    except Exception as e:
+        print(e)
         raise Exception(f"Error: could not set feedback_user_id {user_id} for inference {inference_id}")
     
 
@@ -274,7 +275,7 @@ def get_inference_object(cursor, inference_object_id: str):
                 valid,
                 top_id,
                 upload_date,
-                updated_at
+                update_at
             FROM 
                 object
             WHERE 
@@ -330,7 +331,8 @@ def set_inference_object_top_id(cursor, inference_object_id: str, top_id:str):
             UPDATE 
                 object
             SET
-                top_id = %s
+                top_id = %s,
+                update_at = now()
             WHERE 
                 id = %s
             """
@@ -379,8 +381,7 @@ def set_inference_object_verified_id(cursor, inference_object_id: str, verified_
             UPDATE 
                 object
             SET
-                verified_id = %s,
-                updated_at = CURRENT_TIMESTAMP
+                verified_id = %s
             WHERE 
                 id = %s
             """
