+++ tests/testBin_mass_import.py
