+++ tests/testBin_upload_picture_set.py
